Fixed typo.
Free Software Free Society

To support commits by ctubio,
you can buy-me-a-drink with a small git tip at:
  1GitTipgxvKB3zjCLXRcSgDhC9pivkpc7u

I promise to drink chocolate milk in the development of the next commit.

To request new features or in case this commit breaks something for you,
please create a new github issue with all possible details,
but never share your API Keys!

+++ src/lib/stdev.h
@@ -9,7 +9,7 @@ namespace K {
   using v8::Value;
   using v8::Number;
   using v8::Object;
-  using v8::Float32Array;
+  using v8::Float64Array;
 
   class Stdev {
     public:
